segregated functions according to usecase
